% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrf_rates.R
\name{rrf_rates}
\alias{rrf_rates}
\title{Estimate relative lineage rates.}
\usage{
rrf_rates(phy, outgroup = "", filename = "")
}
\arguments{
\item{phy}{an object of class "phylo".}

\item{outgroup}{a character list containing names of tips belonging to the rooting outgroup, which will be removed in the calculation. If outgroup = "" (the default), the input tree must be rooted and no tips will be removed.}

\item{filename}{a file name specified for the output file.}
}
\value{
A table of relative lineage rates (<filename>_RRF_rates.csv).
}
\description{
This function estimates relative lineage rates from a branch length tree using the relative rate framework (RRF).
}
\examples{
tr <- ape::read.tree("example.nwk")
rrf_rates(phy = tr, outgroup = c("Ornithorhynchus_anatinus", "Zaglossus_bruijni", "Tachyglossus_aculeatus"), filename = "example")
}
\references{
K. Tamura et al. Mol. Biol. Evol. (2018) 35:1770-1782. doi:10.1093/molbev/msy044.
}
\author{
Qiqing Tao (qiqing.tao@temple.edu) and Sudhir Kumar
}
