% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree2table.R
\name{tree2table}
\alias{tree2table}
\title{Convert a NEWICK or NEXUS tree to a table.}
\usage{
tree2table(
  tree.name = "",
  type = c("NEWICK", "NEXUS"),
  time = FALSE,
  filename = ""
)
}
\arguments{
\item{tree.name}{a file name of the branch length tree.}

\item{type}{specify the format of branch length tree. The default is NEWICK.}

\item{time}{logical. If FALSE (default), the input tree is a branch length tree (i.e., not a timetree) and the branch lengths are outputted. If TRUE, the input tree is a timetree and the node times and branch times are outputted.}

\item{filename}{a file name specified for the output files.}
}
\value{
A table of node label, node ID, descendant nodes, branch lengths, and node times (<filename>.csv).
}
\description{
This function converts a tree in NEWICK or NEXUS format to a table (.csv).
}
\examples{
tree2table(tree.name = "example.nwk", type= "NEWICK", filename = "example", time = FALSE)
}
\author{
Qiqing Tao (qiqing.tao@temple.edu) and Sudhir Kumar
}
