% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrf_rates_times.R
\name{rrf_rates_times}
\alias{rrf_rates_times}
\title{Estimate relative lineage rates and node times.}
\usage{
rrf_rates_times(phy, outgroup = "", filename = "", plot = TRUE)
}
\arguments{
\item{phy}{an object of class "phylo".}

\item{outgroup}{a character list containing names of tips belonging to the rooting outgroup, which will be removed in the calculation. If outgroup = "" (the default), the input tree must be rooted and no tips will be removed.}

\item{filename}{a file name specified for the output files.}

\item{plot}{logical. If TRUE (default), the inferred timetree is plotted with the branches colored by the relative rates.}
}
\value{
A table of relative lineage rates and node times (<filename>_RRF_table.csv).

A timetree with branches colored by the relative rates (<filename>_RRF_timetree.nexus).
}
\description{
This function estimates relative lineage rates and node times from a branch length tree using the relative rate framework (RRF).
}
\examples{
tr <- ape::read.tree("example.nwk")
rrf_rates_times(phy = tr, outgroup = c("Ornithorhynchus_anatinus", "Zaglossus_bruijni", "Tachyglossus_aculeatus"), filename = "example", plot = TRUE)
}
\references{
K. Tamura et al. Mol. Biol. Evol. (2018) 35:1770-1782. doi:10.1093/molbev/msy044.
}
\author{
Qiqing Tao (qiqing.tao@temple.edu) and Sudhir Kumar
}
